Create GPUtest.cpp
Added the old gpu codes from the hackathon into the new explicitMatrix code. 

Functionality has changed, so it may not be all correct. e.g. we have a functioning spline interpolator for T and rho,  which was not the case originally. This code does not have the spline interpolator integrated, but it can be added.